+++ R/gg_hbar.R
@@ -324,6 +324,7 @@ gg_hbar <- function(data,
 #' @param y_zero For a numeric y variable, TRUE or FALSE of whether the minimum of the y scale is zero. Defaults to FALSE.
 #' @param y_zero_line For a numeric y variable, TRUE or FALSE of whether to add a zero reference line to the y scale. Defaults to TRUE if there are positive and negative values in y_var. Otherwise defaults to FALSE.   
 #' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles.
+#' @param col_label_digits If numeric colour method, the number of digits to round the labels to. Only applicable where col_labels equals NULL.
 #' @param col_labels A function or named vector to modify colour scale labels. Defaults to stringr::str_to_sentence for categorical colour variables and scales::comma for numeric colour variables. Use ggplot2::waiver() to keep colour labels untransformed.   
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
 #' @param col_na TRUE or FALSE of whether to include col_var NA values. Defaults to TRUE.
@@ -399,6 +400,7 @@ gg_hbar_col <- function(data,
                         y_zero = FALSE,
                         y_zero_line = NULL,
                         col_cuts = NULL,
+                        col_label_digits = NULL,
                         col_labels = NULL,
                         col_method = NULL,
                         col_na = TRUE,
@@ -514,7 +516,7 @@ gg_hbar_col <- function(data,
       })
     }
     
-    if (is.null(col_labels)) col_labels <- scales::comma
+    if (is.null(col_labels)) col_labels <- scales::comma_format(accuracy = 10 ^ -col_label_digits)
     
     if (is.function(col_labels)) {
       data <- data %>% 
@@ -1062,6 +1064,7 @@ gg_hbar_facet <- function(data,
 #' @param y_zero_line For a numeric y variable, TRUE or FALSE of whether to add a zero reference line to the y scale. Defaults to TRUE if there are positive and negative values in y_var. Otherwise defaults to FALSE.   
 #' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles.
 #' @param col_labels A function or named vector to modify colour scale labels. Defaults to stringr::str_to_sentence for categorical colour variables and scales::comma for numeric colour variables. Use ggplot2::waiver() to keep colour labels untransformed.   
+#' @param col_label_digits If numeric colour method, the number of digits to round the labels to. Only applicable where col_labels equals NULL.
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
 #' @param col_na TRUE or FALSE of whether to include col_var NA values. Defaults to TRUE.
 #' @param col_pretty_n For a numeric colour variable of "bin" col_method, the desired number of intervals on the colour scale, as calculated by the pretty algorithm. Defaults to 5. 
@@ -1137,6 +1140,7 @@ gg_hbar_col_facet <- function(data,
                               y_zero_line = NULL,
                               col_cuts = NULL,
                               col_labels = NULL,
+                              col_label_digits = NULL,
                               col_method = NULL,
                               col_na = TRUE,
                               col_pretty_n = 5,
@@ -1264,7 +1268,7 @@ gg_hbar_col_facet <- function(data,
       })
     }
 
-    if (is.null(col_labels)) col_labels <- scales::comma
+    if (is.null(col_labels)) col_labels <- scales::comma_format(accuracy = 10 ^ -col_label_digits)
     
     if (is.function(col_labels)) {
       data <- data %>%