+++ R/gg_point.R
@@ -298,6 +298,7 @@ gg_point <- function(data,
 #' @param y_zero_line For a numeric y variable, TRUE or FALSE whether to add a zero reference line to the y scale. Defaults to TRUE if there are positive and negative values in y_var. Otherwise defaults to FALSE.  
 #' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles.
 #' @param col_labels A function or named vector to modify colour scale labels. Defaults to stringr::str_to_sentence for categorical colour variables and scales::comma for numeric colour variables. Use ggplot2::waiver() to keep colour labels untransformed.   
+#' @param col_label_digits If numeric colour method, the number of digits to round the labels to. Only applicable where col_labels equals NULL.
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
 #' @param col_na TRUE or FALSE of whether to include col_var NA values. Defaults to TRUE.
 #' @param col_pretty_n For a numeric colour variable of "bin" col_method, the desired number of intervals on the colour scale, as calculated by the pretty algorithm. Defaults to 5. 
@@ -356,6 +357,7 @@ gg_point_col <- function(data,
                          col_title = NULL,
                          caption = NULL,
                          col_cuts = NULL,
+                         col_label_digits = NULL,
                          col_labels = NULL,
                          col_method = NULL,
                          col_na = TRUE,
@@ -445,7 +447,7 @@ gg_point_col <- function(data,
       })
     }
     
-    if (is.null(col_labels)) col_labels <- scales::comma
+    if (is.null(col_labels)) col_labels <- scales::comma_format(accuracy = 10 ^ -col_label_digits)
     
     if (is.function(col_labels)) {
       data <- data %>% 
@@ -948,6 +950,7 @@ gg_point_facet <- function(data,
 #' @param y_zero For a numeric y variable, TRUE or FALSE of whether the minimum of the y scale is zero. Defaults to TRUE.
 #' @param y_zero_line For a numeric y variable, TRUE or FALSE whether to add a zero reference line to the y scale. Defaults to TRUE if there are positive and negative values in y_var. Otherwise defaults to FALSE.  
 #' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles. 
+#' @param col_label_digits If numeric colour method, the number of digits to round the labels to. Only applicable where col_labels equals NULL.
 #' @param col_labels A function or named vector to modify colour scale labels. Defaults to stringr::str_to_sentence for categorical colour variables and scales::comma for numeric colour variables. Use ggplot2::waiver() to keep colour labels untransformed.   
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
 #' @param col_na TRUE or FALSE of whether to include col_var NA values. Defaults to TRUE.
@@ -1016,6 +1019,7 @@ gg_point_col_facet <-
            y_zero = FALSE,
            y_zero_line = NULL,
            col_cuts = NULL,
+           col_label_digits = NULL,
            col_labels = NULL,
            col_method = NULL,
            col_na = TRUE,
@@ -1120,7 +1124,7 @@ gg_point_col_facet <-
         })
       }
 
-      if (is.null(col_labels)) col_labels <- scales::comma
+      if (is.null(col_labels)) col_labels <- scales::comma_format(accuracy = 10 ^ -col_label_digits)
       
       if (is.function(col_labels)) {
         data <- data %>%