+++ R/gg_tile.R
@@ -31,6 +31,7 @@
 #' @param y_title y scale title string. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.
 #' @param y_title_wrap Number of characters to wrap the y title to. Defaults to 50. 
 #' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles.
+#' @param col_label_digits If numeric colour method, the number of digits to round the labels to. Only applicable where col_labels equals NULL.
 #' @param col_labels A function or named vector to modify colour scale labels. Defaults to stringr::str_to_sentence for categorical colour variables and scales::comma for numeric colour variables. Use ggplot2::waiver() to keep colour labels untransformed.  
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
 #' @param col_na TRUE or FALSE of whether to include col_var NA values. Defaults to TRUE.
@@ -93,6 +94,7 @@ gg_tile_col <- function(data,
                        y_title = NULL,
                        y_title_wrap = 50,
                        col_cuts = NULL,
+                       col_label_digits = NULL,
                        col_labels = NULL,
                        col_method = NULL,
                        col_na = TRUE,
@@ -208,7 +210,7 @@ gg_tile_col <- function(data,
       })
     }
     
-    if (is.null(col_labels)) col_labels <- scales::comma
+    if (is.null(col_labels)) col_labels <- scales::comma_format(accuracy = 10 ^ -col_label_digits)
     
     if (is.function(col_labels)) {
       data <- data %>% 
@@ -344,6 +346,7 @@ gg_tile_col <- function(data,
 #' @param y_title y scale title string. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.
 #' @param y_title_wrap Number of characters to wrap the y title to. Defaults to 50. 
 #' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles.
+#' @param col_label_digits If numeric colour method, the number of digits to round the labels to. Only applicable where col_labels equals NULL.
 #' @param col_labels A function or named vector to modify colour scale labels. Defaults to stringr::str_to_sentence for categorical colour variables and scales::comma for numeric colour variables. Use ggplot2::waiver() to keep colour labels untransformed.  
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
 #' @param col_na TRUE or FALSE of whether to include col_var NA values. Defaults to TRUE.
@@ -415,6 +418,7 @@ gg_tile_col_facet <- function(data,
                               y_title = NULL,
                               y_title_wrap = 50,
                               col_cuts = NULL,
+                              col_label_digits = NULL,
                               col_labels = NULL,
                               col_method = NULL,
                               col_na = TRUE,
@@ -548,7 +552,7 @@ gg_tile_col_facet <- function(data,
       })
     }
 
-    if (is.null(col_labels)) col_labels <- scales::comma
+    if (is.null(col_labels)) col_labels <- scales::comma_format(accuracy = 10 ^ -col_label_digits)
     
     if (is.function(col_labels)) {
       data <- data %>%